third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/mouse-click-mouseup-listener-update-distribution-crash-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: clickme.createShadowRoot is not a function
crbug.com/417064 minimal reproduction case for hitting assertion check
Click me!
